Move control-flow-related code to ControlFlowChecker.cs
